(#235) Introduce projectile unit
