+++ src/something.cpp
@@ -72,6 +72,7 @@ typedef SSIZE_T ssize_t;
 #include "something_console.cpp"
 #include "something_particles.cpp"
 #include "something_background.cpp"
+#include "something_projectile.cpp"
 #include "something_game.cpp"
 #include "something_main.cpp"
 #include "something_assets.cpp"
